Bugfix 5th graph (#82)
* Change priority to all

* Change output in analytics

+++ backend/pkg/models/project.go
@@ -21,17 +21,17 @@ type Projects struct {
 }
 
 type ProjectAnalytic struct {
-	Id                    int     `json:"Id"`
-	Key                   string  `json:"Key"`
-	Title                 string  `json:"Name"`
-	AllIssuesCount        int     `json:"allIssuesCount"`
-	OpenIssuesCount       int     `json:"openIssuesCount"`
-	CloseIssuesCount      int     `json:"closeIssuesCount"`
-	ResolvedIssuesCount   int     `json:"resolvedIssuesCount"`
-	ReopenedIssuesCount   int     `json:"reopenedIssuesCount"`
-	InProgressIssuesCount int     `json:"progressIssuesCount"`
-	AverageTime           float64 `json:"averageTime"`
-	AverageIssuesCount    float64 `json:"averageIssuesCount"`
+	Id                    int    `json:"Id"`
+	Key                   string `json:"Key"`
+	Title                 string `json:"Name"`
+	AllIssuesCount        int    `json:"allIssuesCount"`
+	OpenIssuesCount       int    `json:"openIssuesCount"`
+	CloseIssuesCount      int    `json:"closeIssuesCount"`
+	ResolvedIssuesCount   int    `json:"resolvedIssuesCount"`
+	ReopenedIssuesCount   int    `json:"reopenedIssuesCount"`
+	InProgressIssuesCount int    `json:"progressIssuesCount"`
+	AverageTime           int    `json:"averageTime"`
+	AverageIssuesCount    int    `json:"averageIssuesCount"`
 }
 
 type PageCount struct {
